Data Validation

+++ api/Controllers/StockController.cs
@@ -30,15 +30,21 @@ namespace api.Controllers
         [HttpGet]
         public async Task<IActionResult> GetAll()
         {
+            if (ModelState.IsValid)
+                return BadRequest();
+
             var stocks = await _stockRepo.GetAllAsync();
             var stockDto = stocks.Select(s => s.ToStockDto());
             return Ok(stocks);
         }
 
 
-        [HttpGet("{id}")]
+        [HttpGet("{id:int}")]
         public async Task<IActionResult> GetById([FromRoute] int id)
         {
+            if (ModelState.IsValid)
+                return BadRequest();
+
             var stock = await _stockRepo.GetByIdAsync(id);
 
             if (stock == null)
@@ -52,6 +58,9 @@ namespace api.Controllers
         [HttpPost]
         public async Task<IActionResult> Create([FromBody] CreateStockRequestDto stockDto)
         {
+            if (ModelState.IsValid)
+                return BadRequest();
+
             var stockModel = stockDto.ToStockFromCreateDTO();
             await _stockRepo.CreateAsync(stockModel);
             return CreatedAtAction(nameof(GetById), new { id = stockModel.ID }, stockModel.ToStockDto());
@@ -59,10 +68,13 @@ namespace api.Controllers
 
 
         [HttpPut]
-        [Route("{id}")]
+        [Route("{id:int}")]
 
         public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateStockRequestDto updateDto)
         {
+            if (ModelState.IsValid)
+                return BadRequest();
+
             var stockModel = await _stockRepo.UpdateAsync(id, updateDto);
 
             if (stockModel == null)
@@ -76,9 +88,12 @@ namespace api.Controllers
 
 
         [HttpDelete]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> Delete([FromRoute] int id)
         {
+            if (ModelState.IsValid)
+                return BadRequest();
+
             var stockModel = await _stockRepo.DeleteAsync(id);
 
             if (stockModel == null)
